make addon_name mandatory, do not attempt to guess addon_dir
Indeed, in some circumstances, pip and/or setuptools copy the
distribution to a temporary directory, so we have no guarantee
that the directory containing setup.py is named after the addon.

Moreover documentation says setup.py should always be called from
the setup.py directory, so '.' is an adequate default for addon_dir
and we don't try to bee too smart with inspect.